Create F1 dataset link pred task (#177)
Create the F1 dataset link pred task which is to predict a list of
constructors a driver will join in the next 10 years, according to F1
race results.
This seems to be one of few link pred task we can create. Tables having
timestamp include `results`, `qualifying`, `standings` and `races`.
Following is the schema of the database

![image](https://github.com/snap-stanford/relbench/assets/21955420/e500b1e9-be8a-4053-b686-f00e978ca755)
Previously @rusty1s proposed to make a link pred task: for each race
predict top k drivers. But race actually has timestamp and it can't be
set as src entities because our current link pred task doesn't support
prediction with new entities (dangling filtered)
So the proposed link pred task is to predict a list of constructors a
driver will join in the next 10 years, according to F1 race results. The
number of entities and degree information is shown below (1/1/2002 is
for the val table and 1/1/2013 is for the test table)

| Timestamps  | Total # src nodes | Total # dst nodes | Avg Degree |
|-------------|------------|------------|-------------|
| 1/14/1952   | 341        | 634        | 1.8592      |
| 1/11/1962   | 172        | 402        | 2.3372      |
| 1/9/1972    | 171        | 407        | 2.3801      |
| 1/6/1982    | 111        | 257        | 2.3153      |
| 1/4/1992    | 98         | 200        | 2.0408      |
| 1/1/2002    | 73         | 142        | 1.9452      |
| 1/1/2013    | 57         | 105        | 1.8421      |

Model performance:
Here is the formatted markdown table:

| Model | Split | Link Prediction Precision | Link Prediction Recall |
Link Prediction MAP |

|-----------|-------------|---------------------------|------------------------|---------------------|
| gnn_link | Validation | 0.049315 | 0.093836 | 0.063090 |
| gnn_link | Test | 0.073684 | 0.244737 | 0.109254 |
| idgnn | Validation | 0.038356 | 0.099087 | 0.097717 |
| idgnn | Test | 0.056140 | 0.151170 | 0.149220 |
| past_visit | Validation | 0.038356 | 0.099087 | 0.078995 |
| past_visit | Test | 0.038596 | 0.111111 | 0.067836 |
| global_popularity | Validation | 0.019178 | 0.044977 | 0.044977 |
| global_popularity | Test | 0.021053 | 0.039474 | 0.039474 |

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>