Remove Extra LEFT JOIN In F1 Task SQLs (#316)
For the F1 tasks, all the SQL's inside `make_table` do a `LEFT JOIN`
with `drivers` table. This join is not necessary since the only column
accessed from the `drivers` is `driverId`, which already exists in other
tables such as `results` and `qualifying`.

I verified that the dataframes produced by `make_table` for all 3 tasks
still match the dataframes when specifying `download=True` inside
`get_task`.