Added driver-race-compete task to rel-f1 dataset (#331)
Added a recommendation task to the rel-f1 dataset called
driver-race-compete. The goal of this recommendation task is to predict
which races a given driver will partake in within the next year.

The main merit of this task is that it is the first recommendation task
for the rel-f1 datasets. This is extremely useful for debugging code
dealing with recommendation tasks, since rel-f1 is by far the smallest
dataset in RelBench and hence the fastest to load.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>